Update module to support pickle loading from Python 3
From this (https://stackoverflow.com/a/11314602/5782416) SO post, we can figure out that pickling in Python 2 is incompatible to be loaded in Python 3. There have been some merges to support moving this project from Python 2 (adding brackets around `print`), so I'm assuming this is a desired fix. Error being fixed is the ascii pickle load error.